Move _jupyter_labextension_paths to voila-gridstack
