--- jupyterlab-gridstack/__init__.py
@@ -1,19 +1,0 @@
-
-import json
-import os.path as osp
-
-from ._version import __version__
-
-HERE = osp.abspath(osp.dirname(__file__))
-
-with open(osp.join(HERE, 'labextension', 'package.json')) as fid:
-    data = json.load(fid)
-
-def _jupyter_labextension_paths():
-    return [{
-        'src': 'labextension',
-        'dest': data['name']
-    }]
-
-
-
