--- jupyterlab-gridstack/_version.py
@@ -1,2 +1,0 @@
-version_info = (0, 1, 0)
-__version__ = ".".join(map(str, version_info))
