+++ voila-gridstack/__init__.py
@@ -7,7 +7,21 @@
 # The full license is in the file LICENSE, distributed with this software.  #
 #############################################################################
 
+import json
+import os.path as osp
 from .server_extension import load_jupyter_server_extension  # noqa
+
+HERE = osp.abspath(osp.dirname(__file__))
+
+with open(osp.join(HERE, 'labextension', 'package.json')) as fid:
+    data = json.load(fid)
+
+
+def _jupyter_labextension_paths():
+    return [{
+        'src': 'labextension',
+        'dest': data['name']
+    }]
 
 
 def _jupyter_nbextension_paths():
